backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==46.0.3
